Switch to aliasing methods
