+++ test/integration/usages/usage_enabled_test.rb
@@ -1,0 +1,9 @@
+require 'test_helper'
+
+class UsageEnabledUsageTest < Minitest::Test
+  include Liquid
+
+  def test_live_usages
+    assert_equal true, Usage.results["Usage is enabled"]
+  end
+end
